modified classify_image.py script to be a more comprehensive test of the model and added new test images
